Abstracting out the templates a bit for easier management.
