+++ src/Gente/data/afflictions.js
@@ -1,47 +1,14 @@
 
 import Affliction from '../components/Affliction';
 import {Repository} from '../Repository'
-
-const TRANSFER_METHODS = {
-	SILIVA: "siliva",
-	BLOOD: "blood",
-	AIR: "air",
-	TOUCH: "touch",
-	MENTAL: "mental"
-};
-
-const STAGES = {
-	ONSET: "onset",
-	ACT: "act",
-	RECOVERY: "recovery",
-	COMPLETE: "complete"
-};
+import templates from '../templates/affliction-templates';
 
 let affictionRepository = {
 	init: function() {
 		let AfflicationRepository = new Repository("Afflications Repository", Affliction);
 
-		AfflicationRepository.define("common-cold", 
-			{
-				name: "Common Cold",
-				isContagious: true,
-				transferMethods: [
-					TRANSFER_METHODS.SILIVA,
-					TRANSFER_METHODS.TOUCH
-				]
-			}
-		);
-		
-		AfflicationRepository.define("flu", 
-			{
-				name: "Flu",
-				isContagious: true,
-				transferMethods: [
-					TRANSFER_METHODS.SILIVA,
-					TRANSFER_METHODS.TOUCH
-				]
-			}
-		);
+		AfflicationRepository.define("common-cold", templates.COMMON_COLD); 
+		AfflicationRepository.define("flu", templates.FLU);
 
 		return AfflicationRepository;
 	}
